[VS-7338] Startup and Program classes combined
